+++ deltafi-ui/src/composables/usePropertySets.ts
@@ -20,7 +20,7 @@ import { ref } from 'vue'
 import useGraphQL from './useGraphQL'
 
 export default function usePropertySets() {
-  const { response, queryGraphQL, loading, loaded, errors } = useGraphQL('config');
+  const { response, queryGraphQL, loading, loaded, errors } = useGraphQL();
   const data = ref(null);
 
   const fetchQuery = {